backend/requirements.txt
# --- Core Framework ---
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic[email]==2.5.3
pydantic-settings==2.1.0

# --- Database (MongoDB & Redis) ---
motor==3.3.2
pymongo==4.6.1
redis==5.0.1

# --- Security & Auth ---
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# --- AI & External Services ---
google-genai==1.2.0    # <--- Swapped OpenAI for Gemini
httpx==0.26.0
python-dotenv==1.0.1

# --- Utilities ---
email-validator==2.1.0
tiktoken==0.5.2
tenacity==8.2.3